Fix php warning
Notice: register_rest_route was called incorrectly. The REST API route definition for site-functionality/v1/rolespermission_callback is missing the required argument. For REST API routes that are intended to be public, use __return_true as the permission callback. Please see Debugging in WordPress for more information. (This message was added in version 5.5.0.) in /var/www/html/wp-includes/functions.php on line 5663